backend/requirements.txt
# Recon Sentinel — Backend Dependencies
# Install: pip install -r requirements.txt

# ─── Core Framework ───────────────────────────────────────────
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic[email]==2.10.*

# ─── Database ─────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*

# ─── Task Queue ───────────────────────────────────────────────
celery[redis]==5.4.*
redis==5.2.*

# ─── Authentication ───────────────────────────────────────────
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
python-multipart==0.0.*

# ─── LLM ──────────────────────────────────────────────────────
litellm==1.57.*
# langgraph — install separately: pip install langgraph

# ─── Utilities ────────────────────────────────────────────────
httpx==0.28.*
python-dotenv==1.0.*
pyyaml==6.0.*

# ─── Report Generation ───────────────────────────────────────
python-docx==1.1.*
# weasyprint — install via system package for PDF support
